python/actions/create-review-report-0.0.0/requirements.txt
fpdf2
ftrack_python_api
ftrack_action_handler